+++ jdaviz/app.py
@@ -26,6 +26,7 @@ from glue.core.state_objects import State
 from glue.core.subset import Subset, RangeSubsetState, RoiSubsetState
 from glue_jupyter.app import JupyterApplication
 from glue_jupyter.state_traitlets_helpers import GlueState
+from glue_jupyter.bqplot.profile import BqplotProfileView
 from ipyvuetify import VuetifyTemplate
 
 from jdaviz.core.config import read_configuration, get_configuration
@@ -617,10 +618,12 @@ class Application(VuetifyTemplate, HubListener):
                     this_type = type(value.subset_state)
 
                 # Skip spatial or spectral subsets if only the other is wanted
-                if subset_type == "spectral" and this_type == RoiSubsetState:
-                    continue
-                elif subset_type == "spatial" and this_type == RangeSubsetState:
-                    continue
+                if subset_type == "spectral" or isinstance(viewer, BqplotProfileView):
+                    if this_type == RoiSubsetState:
+                        continue
+                elif subset_type == "spatial" or not isinstance(viewer, BqplotProfileView):
+                    if this_type == RangeSubsetState:
+                        continue
 
                 # Range selection on a profile is currently not supported in
                 #  the glue translation machinery for astropy regions, so we
@@ -645,6 +648,16 @@ class Application(VuetifyTemplate, HubListener):
                             np.where(value.to_mask() == True)[0], # noqa
                             value.data.coords.spectral_axis)
 
+                    regions[key] = subregions_in_subset
+                    continue
+
+                temp_data = self.get_data_from_viewer(viewer_reference, value.label)
+                if isinstance(temp_data, Spectrum1D):
+                    # Note that we look for mask == False here, rather than True above,
+                    # because specutils masks are the reverse of Glue (of course)
+                    subregions_in_subset = _get_all_subregions(
+                             np.where(~temp_data.mask)[0], # noqa
+                             temp_data.spectral_axis)
                     regions[key] = subregions_in_subset
                     continue